feat: stage 4 - FreeRTOS modem on Core 1, AMP driver, software modem
Software modem (libs/modem, portable C11, no heap): CCSDS K=7
convolutional code with the punctured rates and a soft-decision
Viterbi decoder, Gray BPSK/QPSK/8PSK with max-log demapping, five
MODCODs, RRC pulse shaping, a physical-layer frame with ASM, CRC-4
protected header and pilots, idle PN9 frames, and a streaming receiver
(AGC, Gardner timing with a polyphase interpolator, ASM sync, carrier
PLL, Es/N0 estimate). Every MODCOD decodes at threshold + 1 dB.

Modem application (libs/modem_app): message handling, TX queue, idle
frames, RX forwarding, 1 Hz status and a software loopback. It runs
unchanged in the firmware, in the host simulator and in tests.

FreeRTOS 11.1 firmware for Cortex-A9 Core 1: startup code, an MMU
table built from the address map (only owned regions mapped), GIC,
private-timer tick, console UART, fault capture, AXI DMA
scatter-gather rings, payload_ctrl driver and an event-driven modem
task. Without Linux it runs a self-test through its own IPC rings;
CI boots it in QEMU's Zynq machine.

AMP without remoteproc (ADR-0005, the 6.6 vendor kernel has no Zynq
remoteproc): satlink_amp kernel driver loads the ELF into reserved
memory, takes CPU1 with remove_cpu(), releases it through the SLCR
and the Zynq trampoline, exposes lock-free shared-memory rings as
/dev/satlink-amp with GIC doorbells, and restarts Core 1 after a
fault report or a stalled heartbeat. The ring code is one source for
kernel, firmware and host. Linux side: ModemClient, SimulatedCore1,
satlink-amp tool, device tree with reserved memory, Yocto recipes.

Verified: 190 host tests (ASan/UBSan clean, 94 % line coverage),
QEMU self-test, the three kernel modules link against
linux-xlnx 6.6 at W=1 and the device trees compile (which found a
duplicate can0 label in the stage 2 fragment, fixed). CI jobs added
for the QEMU self-test and the kernel build.

+++ linux/drivers/ccsds_frame_accel/ccsds_frame_accel_drv.c
@@ -290,7 +290,7 @@ static int satlink_fa_probe(struct platform_device *pdev)
 
 	version_reg = readl(fa->regs + FA_REG_VERSION);
 	dev_info(dev, "ccsds_frame_accel v%u.%u ready, irq %d, DMA %s\n",
-		 version_reg >> FA_VERSION_MAJOR_SHIFT, version_reg & FA_VERSION_MINOR_MASK,
+		 version_reg >> FA_VERSION_MAJOR_SHIFT, (u32)(version_reg & FA_VERSION_MINOR_MASK),
 		 fa->irq, (fa->dma_mm2s && fa->dma_s2mm) ? "attached" : "not yet attached");
 
 	return 0;
